Clean up mutable pointer http api
